Convert SCSS from being desktop-first to mobile-first
Although some mobile components, such as the main menu is much
more complicated this way, and this required undoing lots of
stuff for the desktop version, on the whole mobile first is
easier. The Sass has been extensively refactored to move the
responsive media queries nearer to each specific element, and
nesting has been reduced. This is to make it easier to tell
what a given element does and how it changes with sizes.

Hugo min version has been bumped up to 0.129.0, which is unfortunately
no longer compatible w/ 0.111.3, due to unavoidable API changes.